refactor: spliting code for readability
