+++ constants.py
@@ -1,0 +1,9 @@
+from os import environ
+
+CONSTANTS = {
+    'Consumer_Key' : environ['Consumer_Key'],
+    'Consumer_Secret' : environ['Consumer_Secret'],
+    'Access_Key' : environ['Access_Key'],
+    'Access_Secret' : environ['Access_Secret'],
+    'moderator_id' : environ['moderator_id']
+}
